Use 'stdint.h' for more precise control over data types.

+++ include/stereobm_correspondence.h
@@ -4,6 +4,6 @@
 #include "stereobm_image.h"
 
 struct depth_image_t stereobm_correspondence(struct g_image_t *left_image, struct g_image_t *right_image);
-int stereobm_sad_score(struct g_image_t *left_image, struct g_image_t *right_image, int x, int y, int d);
+int32_t stereobm_sad_score(struct g_image_t *left_image, struct g_image_t *right_image, uint16_t x, uint16_t y, uint8_t d);
 
 #endif
